Réorganisation pour plusieurs types de connection
retrait du JDropDownButton et ajout d'une classe SocketConnexion
